backend/pyproject.toml
[project]
name = "miru-backend"
version = "0.1.0"
description = "FastAPI backend for Miru AI Assistant"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.29.0",
    "supabase>=2.3.0",
    "openrouter>=0.7.11",
    "crewai[tools]>=1.10.1",
    "crewai-tools>=0.12.1",
    "tavily-python>=0.5.0",
    "numpy>=1.26.0",
    "pydantic-settings>=2.10.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.28.0",
    "webauthn>=2.7.0",
    "PyJWT[crypto]>=2.10.0",
    "sentry-sdk[fastapi]>=2.1.0",
    "instructor>=1.7.0",
    "authlib>=1.3.0",
    "tortoise-orm>=0.24.0",
    "aerich>=0.7.2",
    "asyncpg>=0.30.0",
    "nest-asyncio",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
    "black>=24.4.0",
    "isort>=5.13.0",
    "mypy>=1.10.0",
    "types-requests>=2.31.0",
    "types-python-dateutil>=2.8.19",
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "safety>=3.2.0",
    "pip-audit>=2.7.0",
]

[tool.setuptools]
packages = ["app"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # Pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "W",   # pycodestyle warnings
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
]
ignore = [
    "E501",   # Line too long (handled by formatter)
    "UP047",  # generic function should use type parameters (not supported in 3.11)
    "UP043",  # unnecessary default type args
    "TC001",  # Move app import into TYPE_CHECKING — breaks FastAPI Depends at runtime
    "TC003",  # Move stdlib import into TYPE_CHECKING — breaks Pydantic at runtime
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.black]
line-length = 100
target-version = ["py313"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["app"]

[tool.mypy]
python_version = "3.13"
exclude = ["build/"]
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
ignore_missing_imports = true
disable_error_code = ["call-arg", "call-overload"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.aerich]
tortoise_orm = "app.infrastructure.database.tortoise.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
